moving fixture realted functions to utils dir
